refactor: Move database (address & chain/tx/cell) modules to subfolder
