+++ packages/neuron-wallet/src/database/address/entities/address.ts
@@ -1,7 +1,7 @@
 import { Entity, BaseEntity, PrimaryColumn, Column, AfterInsert, AfterUpdate, AfterRemove } from 'typeorm'
-import { AddressType } from '../../models/keys/address'
+import { AddressType } from '../../../models/keys/address'
 import { Address as AddressInterface } from '../dao'
-import AddressDbChangedSubject from '../../subjects/address-db-changed-subject'
+import AddressDbChangedSubject from '../../../subjects/address-db-changed-subject'
 
 export enum AddressVersion {
   Testnet = 'testnet',
