+++ packages/neuron-wallet/src/database/chain/entities/input.ts
@@ -1,5 +1,5 @@
 import { Entity, BaseEntity, Column, ManyToOne, PrimaryGeneratedColumn } from 'typeorm'
-import { OutPoint, Input as InputInterface, CellOutPoint } from '../app-types/types'
+import { OutPoint, Input as InputInterface, CellOutPoint } from '../../../app-types/types'
 import Transaction from './transaction'
 
 /* eslint @typescript-eslint/no-unused-vars: "warn" */
