+++ packages/neuron-wallet/src/database/chain/entities/transaction.ts
@@ -1,5 +1,5 @@
 import { Entity, BaseEntity, PrimaryColumn, Column, OneToMany } from 'typeorm'
-import { Witness, OutPoint, Transaction as TransactionInterface } from '../app-types/types'
+import { Witness, OutPoint, Transaction as TransactionInterface } from '../../../app-types/types'
 import InputEntity from './input'
 import OutputEntity from './output'
