+++ packages/neuron-wallet/src/services/addresses.ts
@@ -2,9 +2,9 @@ import { AddressPrefix } from '@nervosnetwork/ckb-sdk-utils'
 import { AccountExtendedPublicKey } from '../models/keys/key'
 import Address, { AddressType } from '../models/keys/address'
 import LockUtils from '../utils/lock-utils'
-import AddressDao, { Address as AddressInterface } from '../addresses/dao'
+import AddressDao, { Address as AddressInterface } from '../database/address/dao'
 import env from '../env'
-import { AddressVersion } from '../addresses/entities/address'
+import { AddressVersion } from '../database/address/entities/address'
 import AddressesUsedSubject from '../subjects/addresses-used-subject'
 
 const MAX_ADDRESS_COUNT = 30
