+++ packages/neuron-wallet/src/services/sync-blocks.ts
@@ -4,8 +4,8 @@ import { Subject, BehaviorSubject } from 'rxjs'
 import Core from '@nervosnetwork/ckb-sdk-core'
 import { Script, OutPoint, Cell, Input, Transaction, Block, BlockHeader } from '../app-types/types'
 import TransactionsService from './transactions'
-import OutputEntity from '../entities/output'
-import SyncInfoEntity from '../entities/sync-info'
+import OutputEntity from '../database/chain/entities/output'
+import SyncInfoEntity from '../database/chain/entities/sync-info'
 import NodeService from './node'
 import LockUtils from '../utils/lock-utils'
 import TypeConvert from '../app-types/type-convert'
