+++ packages/neuron-wallet/src/services/transactions.ts
@@ -2,9 +2,9 @@ import { getConnection } from 'typeorm'
 import { ReplaySubject } from 'rxjs'
 import { OutPoint, Script, Transaction, TransactionWithoutHash, Input, Cell } from '../app-types/types'
 import CellsService from './cells'
-import InputEntity from '../entities/input'
-import OutputEntity from '../entities/output'
-import TransactionEntity from '../entities/transaction'
+import InputEntity from '../database/chain/entities/input'
+import OutputEntity from '../database/chain/entities/output'
+import TransactionEntity from '../database/chain/entities/transaction'
 import NodeService from './node'
 import LockUtils from '../utils/lock-utils'
