+++ packages/neuron-wallet/src/services/wallets.ts
@@ -15,7 +15,7 @@ import ConvertTo from '../app-types/convert-to'
 import Blake2b from '../utils/blake2b'
 import { CurrentWalletNotSet, WalletNotFound, IsRequired, UsedName } from '../exceptions'
 import AddressService from './addresses'
-import { Address as AddressInterface } from '../addresses/dao'
+import { Address as AddressInterface } from '../database/address/dao'
 import Keychain from '../models/keys/keychain'
 import { updateApplicationMenu } from '../utils/application-menu'
 import AddressesUsedSubject from '../subjects/addresses-used-subject'
