+++ packages/neuron-wallet/src/startup/sync-block-task/create.ts
@@ -3,7 +3,7 @@ import { Subject } from 'rxjs'
 import path from 'path'
 import { networkSwitchSubject, NetworkWithID } from '../../services/networks'
 import env from '../../env'
-import initConnection from '../../typeorm'
+import initConnection from '../../database/chain/ormconfig'
 import genesisBlockHash from './genesis'
 import CellsService from '../../services/cells'
 import LockUtils from '../../utils/lock-utils'
